requirements.txt
flask
selenium
mysql-connector-python
